RST-5256 band calculation integration  (#1535)
* Refactoring calculation for children band format
* band in progress
* band calculation paper app
* Paper application income journey
